# apps/server/internal/api/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Robot Center 관제 서버 API입니다. Robot, Operator, Recorder, System 역할별 API를 제공합니다.",
        "title": "Robot Center 관제 API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/operator/robots": {
            "get": {
                "description": "관제 서버에 등록된 로봇 목록을 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "로봇 목록 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "관제 서버에 로봇을 등록하고 로봇 런타임 접속 정보를 함께 반환합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "로봇 생성",
                "parameters": [
                    {
                        "description": "로봇 생성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.CreateRobotRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.CreateRobotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/robots/{robotCode}": {
            "delete": {
                "description": "로봇을 active 목록에서 제외합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "로봇 보관 처리",
                "parameters": [
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotEnvelopeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "로봇 표시 이름과 모델명을 수정합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "로봇 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "로봇 수정 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.UpdateRobotRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotEnvelopeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/robots/{robotCode}/connection-info": {
            "get": {
                "description": "로봇 런타임 접속에 필요한 serverUrl, robotCode, robotToken을 조회합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "로봇 연결 정보 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotConnectionInfoEnvelopeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/operator/robots/{robotCode}/connection-token": {
            "post": {
                "description": "로봇 API용 robotToken을 재발급합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator API"
                ],
                "summary": "로봇 token 재발급",
                "parameters": [
                    {
                        "type": "string",
                        "description": "로봇 코드",
                        "name": "robotCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotConnectionInfoEnvelopeResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/robot/heartbeat": {
            "post": {
                "security": [
                    {
                        "RobotBearerAuth": []
                    }
                ],
                "description": "Robot token으로 인증된 로봇의 heartbeat와 상태를 반영합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot API"
                ],
                "summary": "Robot heartbeat 반영",
                "parameters": [
                    {
                        "description": "Robot heartbeat 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotHeartbeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotHeartbeatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/robot/mission": {
            "get": {
                "security": [
                    {
                        "RobotBearerAuth": []
                    }
                ],
                "description": "Robot token으로 인증된 로봇의 현재 active mission과 SFU 연결 정보를 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Robot API"
                ],
                "summary": "Robot active mission 조회",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotMissionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "robot-center_apps_server_internal_api_dto.CreateRobotRequest": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.CreateRobotResponse": {
            "type": "object",
            "properties": {
                "connectionInfo": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotConnectionInfoResponse"
                },
                "robot": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotResponse"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.RobotConnectionInfoEnvelopeResponse": {
            "type": "object",
            "properties": {
                "connectionInfo": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotConnectionInfoResponse"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotConnectionInfoResponse": {
            "type": "object",
            "properties": {
                "robotCode": {
                    "type": "string"
                },
                "robotToken": {
                    "type": "string"
                },
                "serverUrl": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotEnvelopeResponse": {
            "type": "object",
            "properties": {
                "robot": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotResponse"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotHeartbeatRequest": {
            "type": "object",
            "properties": {
                "batteryPercent": {
                    "type": "integer"
                },
                "networkQuality": {
                    "type": "string"
                },
                "sentAt": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotHeartbeatResponse": {
            "type": "object",
            "properties": {
                "robotCode": {
                    "type": "string"
                },
                "serverTime": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_domain.RobotDeviceState"
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.RobotMissionResponse": {
            "type": "object",
            "properties": {
                "dataChannels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "missionCode": {
                    "type": "string"
                },
                "missionStatus": {
                    "type": "string"
                },
                "serverTime": {
                    "type": "string"
                },
                "sfu": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotSFUConfigResponse"
                },
                "tracks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "turnServers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotTurnServerResponse"
                    }
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastSeenAt": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string"
                },
                "robotCode": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/robot-center_apps_server_internal_domain.RobotConnectionState"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotSFUConfigResponse": {
            "type": "object",
            "properties": {
                "iceTransportPolicy": {
                    "type": "string"
                },
                "signalingUrl": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotTurnServerResponse": {
            "type": "object",
            "properties": {
                "credential": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_api_dto.RobotsResponse": {
            "type": "object",
            "properties": {
                "robots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/robot-center_apps_server_internal_api_dto.RobotResponse"
                    }
                }
            }
        },
...
        "robot-center_apps_server_internal_api_dto.UpdateRobotRequest": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "modelName": {
                    "type": "string"
                }
            }
        },
        "robot-center_apps_server_internal_domain.RobotConnectionState": {
            "type": "string",
            "enum": [
                "fault",
                "offline",
                "online"
            ],
            "x-enum-varnames": [
                "RobotConnectionStateFault",
                "RobotConnectionStateOffline",
                "RobotConnectionStateOnline"
            ]
        },
        "robot-center_apps_server_internal_domain.RobotDeviceState": {
            "type": "string",
            "enum": [
                "fault",
                "offline",
                "online"
            ],
            "x-enum-varnames": [
                "RobotDeviceStateFault",
                "RobotDeviceStateOffline",
                "RobotDeviceStateOnline"
            ]
        },
...
    },
    "securityDefinitions": {
        "RobotBearerAuth": {
            "description": "Robot API에서 사용하는 Bearer robotToken입니다.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}